requirements.txt
sklearn
numpy
pandas
click
pyserial
matplotlib
xarray
meshpy
IPython